+++ test/db/sync/SyncOperations.test.ts
@@ -8,18 +8,13 @@ import {SyncType, SyncOperation} from '@shared/enums';
 import {
   processUpdatesSyncTypePush,
   processCreatesSyncTypePush,
-  processSyncTypePull,
 } from '@services/db/sync/SyncOperations';
 import {insertSyncUpdate} from '@services/db/sync/utils';
-import {insertRows} from '@services/db/functions';
 
 // Constants
 import {apiFunctions} from '@services/db/sync/Constants';
 
-/*
-   Required when importing @services/db/sync/ProcessFunctions
-   Mocking external dependencies
-*/
+/* Mocking external dependencies */
 jest.mock('react-native-fs', () => ({
   DocumentDirectoryPath: '/mocked/document/directory/path',
 }));
@@ -35,45 +30,22 @@ jest.mock('react-native-sqlite-storage', () => ({
 }));
 /**/
 
-// Mocking the Stat Api Class
-jest.mock('@services/api/swagger/Stat', () => ({
-  Stat: jest.fn().mockImplementation(() => ({
-    createCreate: jest.fn(),
-    updateUpdate: jest.fn(),
-    postStat: jest.fn().mockResolvedValue({
-      // Should be the sampleStat var.
-      // Probaly could be if defining mock in test worked...
-      data: [
-        {
-          stat_id: '67f6127d-13cc-4c27-b91f-2b1f83c48eeb',
-          stat_type: 'water',
-          unit: 'ml',
-          timezone: 'UTC',
-          created_at: '2025-01-01T00:00:00.000',
-          updated_at: '2025-01-01T00:01:00.000',
-          user_id: 1,
-          value: 500,
-        },
-      ],
-    }),
-  })),
-}));
-
 jest.mock('@services/db/sync/utils', () => ({
   ...jest.requireActual('@services/db/sync/utils'),
   insertSyncUpdate: jest.fn(),
-  getLastSyncedForTable: jest.fn(),
-  getQueryObjForTable: jest.fn(),
 }));
 
-jest.mock('@services/db/functions', () => ({
-  ...jest.requireActual('@services/db/functions'),
-  insertRows: jest.fn(),
-  updateRows: jest.fn(),
+// Mocking the Stat Api Class
+jest.mock('@services/api/swagger/Stat', () => ({
+  Stat: jest.fn().mockImplementation(() => ({
+    createCreate: jest.fn().mockResolvedValue({status: 201}),
+    updateUpdate: jest.fn().mockResolvedValue({status: 204}),
+  })),
 }));
 
-describe('Sync ProcessFunctions Tests', () => {
+describe('Sync Operation Tests', () => {
   beforeEach(() => {
+    // Clears 'toHaveBeenCalledTimes' cache
     jest.clearAllMocks();
   });
 
@@ -138,29 +110,5 @@ describe('Sync ProcessFunctions Tests', () => {
     );
 
     expect(insertSyncUpdate).toHaveBeenCalledTimes(0);
-  });
-
-  test("processSyncTypePull sync operation 'Creates'", async () => {
-    const tableToSync: dbTables = dbTables.statTable;
-
-    await processSyncTypePull(
-      tableToSync,
-      apiFunctions[tableToSync],
-      SyncOperation.Creates,
-    );
-
-    expect(insertSyncUpdate).toHaveBeenCalledTimes(1);
-    expect(insertSyncUpdate).toHaveBeenCalledWith({
-      last_synced: sampleStat.created_at,
-      sync_operation: SyncOperation.Creates,
-      sync_type: SyncType.Pull,
-      table_name: dbTables.statTable,
-    });
-    expect(insertRows).toHaveBeenCalledTimes(1);
-    expect(insertRows).toHaveBeenCalledWith(
-      dbTables.statTable,
-      [sampleStat],
-      false,
-    );
   });
 });
